functions/requirements.txt
firebase-functions==0.2.0
firebase-admin==6.5.0
google-cloud-aiplatform==1.47.0
google-cloud-documentai==2.23.0
google-cloud-dlp==3.13.0